#9: trim PR comments per project guide
Audited every comment added in this PR against CLAUDE.md (минимум комментариев;
KDoc только для контрактов; не дублировать commit message). Fixes:

- commonMain DeviceKeyPair KDoc was lying ("EC P-256") — Apple actual is 32 random
  bytes. Trimmed to one-line contract about per-install identity persistence.
- appleMain DeviceKeyPair.apple.kt: 4-line WHY block → 2 lines.
- androidMain TrustedDeviceStore.android.kt: 5-line block about DataStore migration
  rationale (commit-message material) → one-line WHY about public keys not needing
  confidentiality.
- appleTest FileServerPairTest: tightened NSUserDefaults isolation note to one line.
- desktopTest FileServerPairTest: moved FileServerPairFailureTest to its own file so
  filename + class name express the split; dropped the "Separate class because…"
  meta-comment. Trimmed the failure-setup WHY to one line.

Pre-existing comments (FileServerRoutes streaming, FileServer resolvedConnectors,
DeviceKeyPair.jvm POSIX null) are untouched and follow the guide as-is.
