+++ composeApp/src/desktopTest/kotlin/com/tubetoast/tether/network/FileServerPairTest.kt
@@ -125,46 +125,3 @@ class FileServerPairTest {
         }
     }
 }
-
-class FileServerPairFailureTest {
-    // Separate class because we deliberately give it a broken store that throws on persist,
-    // which would corrupt @BeforeTest setup in the happy-path class above.
-    private lateinit var keyPairDir: File
-    private lateinit var notADirectory: File
-    private lateinit var server: FileServer
-    private lateinit var client: HttpClient
-    private var port: Int = 0
-
-    @BeforeTest
-    fun setup() {
-        keyPairDir = Files.createTempDirectory("tether-pair-fail-keys").toFile()
-        notADirectory = Files.createTempFile("tether-pair-fail", ".not-a-dir").toFile()
-        val keyPair = DeviceKeyPair(keyPairDir)
-        // Point the store at an existing regular file: configDir.mkdirs() returns false and
-        // writeText on the storage file throws because the parent is not a directory. Same
-        // exception path as a real disk failure (read-only FS, full disk, EACCES).
-        val throwingStore = TrustedDeviceStore(notADirectory)
-        server = FileServer(port = 0, trustedDeviceStore = throwingStore, deviceKeyPair = keyPair)
-        port = server.start()
-        client = HttpClient(CIO) { install(ContentNegotiation) { json() } }
-    }
-
-    @AfterTest
-    fun teardown() {
-        client.close()
-        server.stop()
-        keyPairDir.deleteRecursively()
-        notADirectory.delete()
-    }
-
-    @Test
-    fun `pair returns 500 when store fails to persist`() {
-        runBlocking {
-            val response = client.post("http://localhost:$port/pair") {
-                contentType(ContentType.Application.Json)
-                setBody(PairRequest(publicKey = byteArrayOf(7, 7, 7), deviceName = "Whatever"))
-            }
-            assertEquals(HttpStatusCode.InternalServerError, response.status)
-        }
-    }
-}
